refactor: import external should not create needless external module
