+++ crates/rspack_plugin_library/src/lib.rs
@@ -11,8 +11,6 @@ mod utils;
 pub use amd_library_plugin::AmdLibraryPlugin;
 pub use assign_library_plugin::*;
 pub use export_property_library_plugin::ExportPropertyLibraryPlugin;
-pub use modern_module::ModernModuleImportDependencyTemplate;
-pub use modern_module_library_plugin::replace_import_dependencies_for_external_modules;
 use rspack_core::{BoxPlugin, PluginExt};
 pub use system_library_plugin::SystemLibraryPlugin;
 pub use umd_library_plugin::UmdLibraryPlugin;
