+++ crates/rspack_plugin_library/src/modern_module/mod.rs
@@ -1,5 +1,2 @@
-mod import_dependency;
 mod reexport_star_external_dependency;
-
-pub use import_dependency::*;
 pub use reexport_star_external_dependency::*;
