+++ crates/rspack_plugin_rslib/src/lib.rs
@@ -1,4 +1,5 @@
+mod import_dependency;
+mod import_external;
 mod parser_plugin;
 mod plugin;
-
 pub use plugin::*;
