+++ crates/rspack_plugin_rslib/src/plugin.rs
@@ -12,16 +12,16 @@ use rspack_hook::{plugin, plugin_hook};
 use rspack_plugin_javascript::{
   BoxJavascriptParserPlugin, parser_and_generator::JavaScriptParserAndGenerator,
 };
-use rspack_plugin_library::{
-  ModernModuleImportDependencyTemplate, replace_import_dependencies_for_external_modules,
-};
 
-use crate::parser_plugin::RslibParserPlugin;
+use crate::{
+  import_dependency::RslibDependencyTemplate,
+  import_external::replace_import_dependencies_for_external_modules,
+  parser_plugin::RslibParserPlugin,
+};
 
 #[derive(Debug)]
 pub struct RslibPluginOptions {
   pub intercept_api_plugin: bool,
-  pub compact_external_module_dynamic_import: bool,
   pub force_node_shims: bool,
 }
 
@@ -76,15 +76,15 @@ async fn compilation(
   _params: &mut CompilationParams,
 ) -> Result<()> {
   compilation.set_dependency_template(
-    ModernModuleImportDependencyTemplate::template_type(),
-    Arc::new(ModernModuleImportDependencyTemplate::default()),
+    RslibDependencyTemplate::template_type(),
+    Arc::new(RslibDependencyTemplate::default()),
   );
   Ok(())
 }
 
 #[plugin_hook(CompilerFinishMake for RslibPlugin)]
 async fn finish_make(&self, compilation: &mut Compilation) -> Result<()> {
-  // Replace ImportDependency instances with ModernModuleImportDependency for external modules
+  // Replace ImportDependency instances with RslibImportDependency for external modules
   replace_import_dependencies_for_external_modules(compilation)?;
   Ok(())
 }
@@ -101,9 +101,7 @@ impl Plugin for RslibPlugin {
       .parser
       .tap(nmf_parser::new(self));
 
-    if self.options.compact_external_module_dynamic_import {
-      ctx.compiler_hooks.finish_make.tap(finish_make::new(self));
-    }
+    ctx.compiler_hooks.finish_make.tap(finish_make::new(self));
 
     Ok(())
   }
